Renamed net.mediatype to net.header.
- No code changes, only package renaming.
