+++ src/test/java/walkingkooka/net/header/MediaTypeParameterNameTest.java
@@ -16,7 +16,7 @@
  *
  */
 
-package walkingkooka.net.media;
+package walkingkooka.net.header;
 
 import org.junit.Test;
 import walkingkooka.naming.NameTestCase;
