+++ src/test/java/walkingkooka/net/header/MediaTypeParserTest.java
@@ -16,23 +16,14 @@
  *
  */
 
-package walkingkooka.net.media;
+package walkingkooka.net.header;
 
-final class OneMediaTypeParser extends MediaTypeParser {
+import walkingkooka.test.PackagePrivateClassTestCase;
 
-    static MediaType parseOneOrFail(final String text) {
-        return new OneMediaTypeParser(text).parse(MODE_TYPE);
-    }
+public final class MediaTypeParserTest extends PackagePrivateClassTestCase<MediaTypeParser> {
 
-    private OneMediaTypeParser(final String text) {
-        super(text);
-    }
-
-    /**
-     * If a comma is encountered when parsing a single mime type it is invalid.
-     */
     @Override
-    void mediaTypeSeparator() {
-        this.failInvalidCharacter();
+    protected Class<MediaTypeParser> type() {
+        return MediaTypeParser.class;
     }
 }
