+++ src/test/java/walkingkooka/net/header/MediaTypeParserTestCase.java
@@ -16,7 +16,7 @@
  *
  */
 
-package walkingkooka.net.media;
+package walkingkooka.net.header;
 
 import org.junit.Test;
 import walkingkooka.collect.map.Maps;
