+++ src/test/java/walkingkooka/net/http/HttpHeaderValueManyMediaTypeConverterTest.java
@@ -21,8 +21,8 @@ package walkingkooka.net.http;
 import org.junit.Test;
 import walkingkooka.collect.list.Lists;
 import walkingkooka.collect.map.Maps;
-import walkingkooka.net.media.MediaType;
-import walkingkooka.net.media.MediaTypeParameterName;
+import walkingkooka.net.header.MediaType;
+import walkingkooka.net.header.MediaTypeParameterName;
 
 import java.util.List;
